jupyter listens on 0; symlink is forced

+++ stage4/06-jupyter/00-run.sh
@@ -13,12 +13,15 @@ cp files/* "${ROOTFS_DIR}/home/${FIRST_USER_NAME}/Notebooks/"
 
 cp ohio.jpg "${ROOTFS_DIR}/usr/share/rpd-wallpaper/temple.jpg"
 
-rm "${ROOTFS_DIR}/etc/xdg/autostart/piwiz.desktop"
+if [[ -f "${ROOTFS_DIR}/etc/xdg/autostart/piwiz.desktop" ]]; then
+	rm "${ROOTFS_DIR}/etc/xdg/autostart/piwiz.desktop"
+fi
 
 cp jupyter.service "${ROOTFS_DIR}/lib/systemd/system/jupyter.service"
-ln -sv "${ROOTFS_DIR}/lib/systemd/system/jupyter.service" "$ROOTFS_DIR/etc/systemd/system/multi-user.target.wants/jupyter.service"
+ln -sfv "${ROOTFS_DIR}/lib/systemd/system/jupyter.service" "$ROOTFS_DIR/etc/systemd/system/multi-user.target.wants/jupyter.service"
 
 on_chroot << EOF
 pip3 install --upgrade matplotlib
 pip3 install --upgrade pillow
+pip3 install --upgrade blinkt
 EOF
